requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.2.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
clip=1.0=pypi_0
cryptography=39.0.1=py37h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.34=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
diffusers=0.21.2=pypi_0
easydict=1.10=pypi_0
einops=0.6.0=pypi_0
faiss-gpu=1.7.3=py3.7_h28a55e0_0_cuda11.3
ffmpeg=4.3=hf484d3e_0
filelock=3.11.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
ftfy=6.1.1=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.53.0=pypi_0
h5py=3.7.0=py37h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.16.4=pypi_0
idna=3.4=py37h06a4308_0
ignite=0.1.2=py37_0
imageio=2.27.0=pypi_0
importlib-metadata=6.3.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
invisible-watermark=0.2.0=pypi_0
ipython=7.31.1=py37h06a4308_1
jedi=0.18.1=py37h06a4308_1
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libfaiss=1.7.3=hfc2d529_0_cuda11.3
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvmlite=0.39.1=pypi_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py37h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
ninja=1.11.1=pypi_0
numba=0.56.4=pypi_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
nvidia-ml-py=12.535.108=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.16.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
prettytable=3.5.0=py37h06a4308_0
prompt-toolkit=3.0.36=py37h06a4308_0
protobuf=3.20.1=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pytorch=1.13.1=py3.7_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-lightning=1.6.5=pypi_0
pytorch-mutex=1.0=cuda
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.3=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
sentencepiece=0.1.98=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.6.13=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch-fidelity=0.3.0=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.14.1=py37_cu117
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py37h06a4308_0
transformers=4.28.1=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
yarl=1.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
